# src/docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "API Documentation for Snicks, the ecommerce store for mens sneakers at [http://swagger.io](http://swagger.io) or on [irc.freenode.net, #swagger](http://swagger.io/irc/).      For this sample, you can use the api key `special-key` to test the authorization     filters.",
    "version": "1.0.0",
    "title": "API Documentation for Snicks",
    "contact": {
      "email": "stepheng323@gmail.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "localhost:4000",
  "basePath": "/api/v1",
...
  "schemes": ["http", "https"],
  "securityDefinitions": {
    "BearerToken": {
      "type": "apiKey",
      "description": "All protected routes must be accessed with a Bearer Token.The JWT Token is generated by the API upon user login.",
      "name": "Authorization",
      "in": "header"
    }
  },
  "paths": {
    "/user": {
      "post": {
        "tags": ["user"],
        "summary": "Create user",
        "description": "A person can creat an account.",
        "operationId": "createUser",
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Created user object",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CreateUser"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/UserLoginResponse"
            }
          },
          "409": {
            "description": "Conflict",
            "schema": {
              "$ref": "#/definitions/UserSignupConflict"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/BadRequest"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/ServerError"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "UserLoginResponse": {
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean",
          "example": true
        },
        "message": {
          "type": "string",
          "example": "User logged in successfully"
        },
        "payload": {
          "type": "object",
          "properties": {
            "id": {
              "type": "integer",
              "example": 1
            },
            "email": {
              "type": "string",
              "example": "stepheng323@gmail.com"
            },
            "firstName": {
              "type": "string",
              "example": "Jon"
            },
            "lastName": {
              "type": "string",
              "example": "Doe"
            },
            "phone": {
              "type": "string",
              "example": "09239944955"
            },
            "role": {
              "type": "string",
              "example": "user"
            },
            "token": {
              "type": "string",
              "example": "eejffioffhbdbfdkfhfdhlfkd.fdbhdhfldhfyofhbdfdfvgdvfgdkjf.fdgkvfvgdfvdjhfldogvkdsfgdls"
            },
            "tokenExpiry": {
              "type": "integer",
              "example": 2873844984
            }
          }
        }
      }
    },
...
    "CreateUser": {
      "type": "object",
      "properties": {
        "firstName": {
          "type": "string",
          "example": "John"
        },
        "lastName": {
          "type": "string",
          "example": "Doe"
        },
        "email": {
          "type": "string",
          "example": "stepheng323@gmail.com"
        },
        "password": {
          "type": "string",
          "example": "12345"
        },
        "phone": {
          "type": "string",
          "example": "080888348"
        },
        "role": {
          "type": "string",
          "example": "user"
        },
        "token": {
          "type": "string",
          "example": "eejffioffhbdbfdkfhfdhlfkd.fdbhdhfldhfyofhbdfdfvgdvfgdkjf.fdgkvfvgdfvdjhfldogvkdsfgdls"
        },
        "tokenExpiry": {
          "type": "integer",
          "example": 2873844984
        }
      },
      "xml": {
        "name": "User"
      }
    },
    "UserSignupConflict": {
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean",
          "example": false
        },
        "message": {
          "type": "string",
          "example": "a user with the email exist"
        },
        "payload": {
          "type": "object"
        }
      }
    },
...
    "BadRequest": {
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean",
          "example": false
        },
        "message": {
          "type": "string",
          "example": "firstname is required"
        }
      }
    },
    "ServerError": {
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean",
          "example": false
        },
        "message": {
          "type": "string",
          "example": "Internal Server Error"
        }
      }
    },
...
  },
...
}